requirement.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///C:/ci/entrypoints_1649926676279/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
ipykernel @ file:///C:/ci/ipykernel_1647000977835/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651602003142/work
jedi @ file:///C:/ci/jedi_1644315428305/work
joblib==1.1.0
jupyter-client @ file:///C:/ci/jupyter_client_1650534191185/work
jupyter-core @ file:///C:/ci/jupyter_core_1651638076360/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848132512/work
nltk==3.7
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py4j==0.10.9.5
pyenchant==3.2.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyspark==3.3.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pywin32==302
pyzmq @ file:///C:/ci_310/pyzmq_1642090360481/work
regex==2022.8.17
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tornado @ file:///C:/ci_310/tornado_1642093111997/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
wincertstore==0.2
